refactor all details option into one file
