--- clients/apps/web/src/components/Benefit/Configuration/BenefitConfigurationProps.ts
@@ -1,6 +1,0 @@
-import { schemas } from '@polar-sh/client'
-
-export interface BenefitConfigurationProps {
-  benefit: schemas['Benefit']
-  organization: schemas['Organization']
-}
